move mi tests from test_loc.py
